+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- Syntax highlighting for SystemVerilog keywords.
 
 ## [0.1.0]
 ### Added